+++ packages/dataviews/src/components/dataform-controls/date.tsx
@@ -1,4 +1,3 @@
-import clsx from 'clsx';
 import {
 	format,
 	isValid as isValidDate,
@@ -11,8 +10,7 @@ import {
 import {
 	BaseControl,
 	Button,
-	Icon as WCIcon,
-	__experimentalInputControl as InputControl,
+	__experimentalInputControl as WCInputControl,
 } from '@wordpress/components';
 import { speak } from '@wordpress/a11y';
 import {
@@ -24,8 +22,12 @@ import {
 } from '@wordpress/element';
 import { __ } from '@wordpress/i18n';
 import { getDate, getSettings } from '@wordpress/date';
-import { error as errorIcon } from '@wordpress/icons';
-import { Calendar, RangeCalendar, Stack } from '@wordpress/ui';
+import {
+	Calendar,
+	RangeCalendar,
+	Stack,
+	ValidityIndicator,
+} from '@wordpress/ui';
 import RelativeDateControl from './utils/relative-date-control';
 import useDisabledDateMatchers from './utils/use-disabled-date-matchers';
 import {
@@ -253,22 +255,10 @@ function ValidatedDateControl< Item >( {
 		<div onBlur={ onBlur }>
 			{ children }
 			{ customValidity && (
-				<p
-					className={ clsx(
-						'components-validated-control__indicator',
-						customValidity.type === 'invalid'
-							? 'is-invalid'
-							: undefined
-					) }
-				>
-					<WCIcon
-						className="components-validated-control__indicator-icon"
-						icon={ errorIcon }
-						size={ 16 }
-						fill="currentColor"
-					/>
-					{ customValidity.message }
-				</p>
+				<ValidityIndicator
+					type={ customValidity.type }
+					message={ customValidity.message }
+				/>
 			) }
 		</div>
 	);
@@ -425,7 +415,7 @@ function CalendarDateControl< Item >( {
 					</Stack>
 
 					{ /* Manual date input */ }
-					<InputControl
+					<WCInputControl
 						ref={ validityTargetRef }
 						type="date"
 						label={ __( 'Date' ) }
@@ -661,7 +651,7 @@ function CalendarDateRangeControl< Item >( {
 						justify="space-between"
 						className="dataviews-controls__date-range-inputs"
 					>
-						<InputControl
+						<WCInputControl
 							ref={ fromInputRef }
 							type="date"
 							label={ __( 'From' ) }
@@ -675,7 +665,7 @@ function CalendarDateRangeControl< Item >( {
 							min={ minConstraint }
 							max={ maxConstraint }
 						/>
-						<InputControl
+						<WCInputControl
 							ref={ toInputRef }
 							type="date"
 							label={ __( 'To' ) }